util_python_lib/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "rhit-bhuwalk",email = "kushbhuwalka@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "zeroentropy (>=0.1.0a3,<0.2.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "playwright (>=1.50.0,<2.0.0)",
    "agentql (>=1.9.0,<2.0.0)",
    "browserbase (>=1.2.0,<2.0.0)",
    "mvt (>=2.6.0,<3.0.0)",
    "mapbox-vector-tile (>=2.1.0,<3.0.0)",
    "shapely (>=2.0.7,<3.0.0)",
    "supabase (>=2.15.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
